\name{negbin1-package}
\alias{negbin1-package}
\alias{negbin1}
\docType{package}
\title{
\packageTitle{negbin1}
}
\description{
\packageDescription{negbin1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{negbin1}
\packageIndices{negbin1}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{negbin1}

Maintainer: \packageMaintainer{negbin1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
